+++ src/neuroconv/tools/spikeinterface/spikeinterface.py
@@ -13,7 +13,7 @@ from spikeinterface import BaseRecording, BaseSorting
 from spikeinterface.core.old_api_utils import OldToNewRecording, OldToNewSorting
 from spikeextractors import RecordingExtractor, SortingExtractor
 from numbers import Real
-from hdmf.data_utils import DataChunkIterator
+from hdmf.data_utils import DataChunkIterator, AbstractDataChunkIterator
 from hdmf.backends.hdf5.h5_utils import H5DataIO
 import psutil
 
@@ -349,7 +349,7 @@ def add_electrodes(
             group_name_array = data_to_add["group"]["data"].astype("str", copy=False)
         else:
             default_group_name = "default_group"
-            group_name_array = np.full_like(channel_ids, fill_value=default_group_name)
+            group_name_array = np.full_like(channel_name_array, fill_value=default_group_name)
 
     group_name_array[group_name_array == ""] = "default_group"
     data_to_add["group_name"].update(description="group_name", data=group_name_array, index=False)
@@ -488,6 +488,68 @@ def check_if_recording_traces_fit_into_memory(recording: SpikeInterfaceRecording
         raise MemoryError(message)
 
 
+def _recording_traces_to_hdmf_iterator(
+    recording: BaseRecording,
+    segment_index: int = None,
+    return_scaled: bool = False,
+    iterator_type: str = "v2",
+    iterator_opts: dict = None,
+) -> AbstractDataChunkIterator:
+    """Function to wrap traces of spikeinterface recording into an AbstractDataChunkIterator.
+
+    Parameters
+    ----------
+    recording : BaseRecording
+        A recording extractor from spikeinterface
+    segment_index : int, optional
+        The recording segment to add to the NWBFile.
+    return_scaled : bool, defaults to False
+        When True recording extractor objects from spikeinterface return their traces in microvolts.
+    iterator_type: str (optional, defaults to 'v2')
+        The type of DataChunkIterator to use.
+        'v1' is the original DataChunkIterator of the hdmf data_utils.
+        'v2' is the locally developed SpikeInterfaceRecordingDataChunkIterator, which offers full control over chunking.
+    iterator_opts: dict (optional)
+        Dictionary of options for the iterator.
+        See https://hdmf.readthedocs.io/en/stable/hdmf.data_utils.html#hdmf.data_utils.GenericDataChunkIterator
+        for the full list of options.
+
+    Returns
+    -------
+    traces_as_iterator: AbstractDataChunkIterator
+        The traces of the recording extractor wrapped in an iterator object.
+
+    Raises
+    ------
+    ValueError
+        If the iterator_type is not 'v1', 'v2' or None.
+    """
+
+    supported_iterator_types = ["v1", "v2", None]
+    if iterator_type not in supported_iterator_types:
+        message = f"iterator_type {iterator_type} should be either 'v1', 'v2' (recommended) or None"
+        raise ValueError(message)
+
+    iterator_opts = dict() if iterator_opts is None else iterator_opts
+
+    if iterator_type is None:
+        check_if_recording_traces_fit_into_memory(recording=recording, segment_index=segment_index)
+        traces_as_iterator = recording.get_traces(return_scaled=return_scaled, segment_index=segment_index)
+    elif iterator_type == "v2":
+        traces_as_iterator = SpikeInterfaceRecordingDataChunkIterator(
+            recording=recording,
+            segment_index=segment_index,
+            return_scaled=return_scaled,
+            **iterator_opts,
+        )
+    elif iterator_type == "v1":
+        traces_as_iterator = DataChunkIterator(
+            data=recording.get_traces(return_scaled=return_scaled, segment_index=segment_index), **iterator_opts
+        )
+
+    return traces_as_iterator
+
+
 def add_electrical_series(
     recording: SpikeInterfaceRecording,
     nwbfile: pynwb.NWBFile,
@@ -509,6 +571,7 @@ def add_electrical_series(
     Parameters
     ----------
     recording: SpikeInterfaceRecording
+        A recording extractor from spikeinterface
     nwbfile: NWBFile
         nwb file to which the recording information is to be added
     metadata: dict
@@ -540,15 +603,12 @@ def add_electrical_series(
     iterator_type: str (optional, defaults to 'v2')
         The type of DataChunkIterator to use.
         'v1' is the original DataChunkIterator of the hdmf data_utils.
-        'v2' is the locally developed RecordingExtractorDataChunkIterator, which offers full control over chunking.
+        'v2' is the locally developed SpikeInterfaceRecordingDataChunkIterator, which offers full control over chunking.
     iterator_opts: dict (optional)
-        Dictionary of options for the RecordingExtractorDataChunkIterator (iterator_type='v2').
-        Valid options are
-            buffer_gb: float (optional, defaults to 1 GB)
-                Recommended to be as much free RAM as available. Automatically calculates suitable buffer shape.
-            chunk_mb: float (optional, defaults to 1 MB)
-                Should be below 1 MB. Automatically calculates suitable chunk shape.
-        If manual specification of buffer_shape and chunk_shape are desired, these may be specified as well.
+        Dictionary of options for the iterator.
+        See https://hdmf.readthedocs.io/en/stable/hdmf.data_utils.html#hdmf.data_utils.GenericDataChunkIterator
+        for the full list of options.
+
     Missing keys in an element of metadata['Ecephys']['ElectrodeGroup'] will be auto-populated with defaults
     whenever possible.
     """
@@ -570,7 +630,7 @@ def add_electrical_series(
     default_description = dict(raw="Raw acquired data", lfp="Processed data - LFP", processed="Processed data")
     eseries_kwargs = dict(name=default_name, description=default_description[write_as])
 
-    # Write as functionality
+    # Select and/or create module if lfp or processed data is to be stored.
     if write_as in ["lfp", "processed"]:
         ecephys_mod = get_module(
             nwbfile=nwbfile,
@@ -602,7 +662,7 @@ def add_electrical_series(
     )
     eseries_kwargs.update(electrodes=electrode_table_region)
 
-    # Spikeinterface guarantee data in micro volts when return_scaled=True. This multiplies by gain and adds offsets
+    # Spikeinterface guarantees data in micro volts when return_scaled=True. This multiplies by gain and adds offsets
     # In nwb to get traces in Volts we take data*channel_conversion*conversion + offset
     channel_conversion = checked_recording.get_channel_gains()
     channel_offset = checked_recording.get_channel_offsets()
@@ -620,31 +680,15 @@ def add_electrical_series(
         eseries_kwargs.update(offset=unique_offset * micro_to_volts_conversion_factor)
 
     # Iterator
-    iterator_opts = dict() if iterator_opts is None else iterator_opts
-
-    if iterator_type is None:
-        check_if_recording_traces_fit_into_memory(recording=checked_recording, segment_index=segment_index)
-        ephys_data = checked_recording.get_traces(return_scaled=write_scaled, segment_index=segment_index)
-
-    elif iterator_type == "v2":
-        ephys_data = SpikeInterfaceRecordingDataChunkIterator(
-            recording=checked_recording,
-            segment_index=segment_index,
-            return_scaled=write_scaled,
-            **iterator_opts,
-        )
-    elif iterator_type == "v1":
-        if isinstance(checked_recording.get_traces(end_frame=5, return_scaled=write_scaled), np.memmap) and np.all(
-            channel_offset == 0
-        ):
-            ephys_data = DataChunkIterator(
-                data=checked_recording.get_traces(return_scaled=write_scaled), **iterator_opts
-            )
-        else:
-            raise ValueError("iterator_type='v1' only supports memmapable trace types! Use iterator_type='v2' instead.")
-    else:
-        raise NotImplementedError(f"iterator_type ({iterator_type}) should be either 'v1' or 'v2' (recommended)!")
-    eseries_kwargs.update(data=H5DataIO(data=ephys_data, compression=compression, compression_opts=compression_opts))
+    ephys_data_iterator = _recording_traces_to_hdmf_iterator(
+        recording=checked_recording,
+        segment_index=segment_index,
+        iterator_type=iterator_type,
+        iterator_opts=iterator_opts,
+    )
+    eseries_kwargs.update(
+        data=H5DataIO(data=ephys_data_iterator, compression=compression, compression_opts=compression_opts)
+    )
 
     # Timestamps vs rate
     timestamps = checked_recording.get_times(segment_index=segment_index)